Add MenuHider easing

+++ src/Home/Menu/Menu.cpp
@@ -37,7 +37,7 @@ void Menu::setItems(std::vector<MenuItem>& items){
 	setOffsetY(64); //used for hiding the menu
 }
 
-void Menu::setOffsetY(uint8_t y){
+void Menu::setOffsetY(int16_t y){
 	offsetY = y;
 	for(auto& item : items){
 		item.image.setY(originY + offsetY);
