+++ src/Home/Menu/Menu.h
@@ -52,7 +52,7 @@ public:
 	void repos();
 	uint getSelectedIndex() const;
 
-	void setOffsetY(uint8_t y);
+	void setOffsetY(int16_t y);
 
 	void shake();
 	bool isShaking();
@@ -62,7 +62,7 @@ private:
 	const uint8_t width = 32;
 	const uint8_t origin;
 	const uint8_t originY = 75;
-	uint8_t offsetY = 0;
+	int16_t offsetY = 0;
 	const float speed = 250;
 	fs::File borderFile;
 	const char* borderPath = "/MenuIcons/Border.raw";
